requirements.txt
asgiref==3.4.1
autopep8 @ file:///opt/conda/conda-bld/autopep8_1639166893812/work
backcall==0.2.0
bcrypt==3.2.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
coreapi==2.3.3
coreschema==0.0.4
decorator==5.1.1
Django==3.2.9
django-cors-headers==3.10.1
django-rest-swagger==2.2.0
djangorestframework==3.13.1
drf-yasg==1.20.0
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
inflection==0.5.1
ipython==7.31.0
itypes==1.2.0
jedi==0.18.1
Jinja2==3.0.3
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
openapi-codec==1.3.2
packaging==21.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
prompt-toolkit==3.0.24
psycopg2 @ file:///tmp/build/80754af9/psycopg2_1612298595717/work
psycopg2-binary==2.9.2
ptyprocess==0.7.0
pycodestyle @ file:///tmp/build/80754af9/pycodestyle_1636635402688/work
pycparser==2.21
Pygments==2.11.2
pyparsing==3.0.9
pytz==2021.3
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
simplejson==3.17.6
SQLAlchemy==1.4.29
sqlparse==0.4.2
toml @ file:///tmp/build/80754af9/toml_1616166611790/work
traitlets==5.1.1
uritemplate==4.1.1
urllib3==1.26.8
wcwidth==0.2.5
whitenoise==5.3.0
